requirements.txt
streamlit
openai==0.28.1
chromadb
PyPDF2
python-dotenv